Change to int to match nearby test + re-run CI

+++ test/backend/test_renderer_failures.py
@@ -85,9 +85,9 @@ class TestWGSLFailures(unittest.TestCase):
 
   # WGSL has a specific select(alt, val, gate) ternary operator instead of gate?val:alt
   def test_gated_load(self):
-    a = UOp.param(0, dtypes.float32.ptr(4))
-    b = UOp.param(1, dtypes.float32.ptr(4))
-    c = UOp.param(2, dtypes.float32.ptr(4))
+    a = UOp.param(0, dtypes.int.ptr(4))
+    b = UOp.param(1, dtypes.int.ptr(4))
+    c = UOp.param(2, dtypes.int.ptr(4))
     lidx0 = UOp(Ops.SPECIAL, dtypes.int, (UOp.const(dtypes.int, 4),), "lidx0")
     gate = lidx0.ne(0)
     alt = c.index(lidx0, ptr=True).load()
@@ -95,7 +95,7 @@ class TestWGSLFailures(unittest.TestCase):
     alt_load = gate.where(ld, alt)
     store = UOp.store(a.index(lidx0, ptr=True), alt_load)
     sink = UOp(Ops.SINK, dtypes.void, (store,), arg=KernelInfo())
-    ret = _test_uop_result([Tensor([0,1,2,3], dtype=dtypes.float32), Tensor([4,5,6,7], dtype=dtypes.float32)], sink, local_size=[4])[0]
+    ret = _test_uop_result([Tensor([0,1,2,3], dtype=dtypes.int), Tensor([4,5,6,7], dtype=dtypes.int)], sink, local_size=[4])[0]
     np.testing.assert_equal(ret, [4,1,2,3])
 
 @unittest.skipIf(not isinstance(Device[Device.DEFAULT].renderer, PTXRenderer), "tests for ptx renderer")
